+++ CHANGELOG.md
@@ -8,7 +8,9 @@ This is used for recording Invader's changes. This changelog is based on [Keep a
 - invader-bitmap: Double multiply sprites now simply replaces the pixel like tool.exe rather than alpha blend into gray
 - invader-bitmap: Usage and the p8 compression flag are now preserved
 - invader-bitmap: Height maps now generate bump maps similar to tool.exe
-- invader-bitmap: Added --usage (default or bumpmap) and --bump-height parameters
+- invader-bitmap: Added --usage (default, bumpmap, and detail) and --bump-height parameters
+- invader-bitmap: --detail-fade replaces --mipmap-fade and is now only usable on detail maps.
+- invader-bitmap: --detail-fade now approximately matches how tool.exe does fade to gray
 - invader-bitmap: Added p8-bump support based on Stubbs the Zombie's palette
 - invader-bitmap: Dithering is now available for 16-bit and palettized bitmaps
 - invader-bitmap: Dithering now takes an argument: `<channels>`. Channels are letters (i.e. `argb`).